Rename engine package to interfaces
This makes the packages more generic, allowing for a restructuring of
the packages.
This includes moving EntityToGraphicsMapper to the interfaces package.
