+++ Core/src/main/java/com/github/sef24sp4/core/interfaces/IGameProcessor.java
@@ -1,4 +1,4 @@
-package com.github.sef24sp4.core.engine;
+package com.github.sef24sp4.core.interfaces;
 
 public interface IGameProcessor {
 	public void begin();
