+++ Core/src/main/java/com/github/sef24sp4/core/interfaces/IGameTickExecutor.java
@@ -1,4 +1,4 @@
-package com.github.sef24sp4.core.engine;
+package com.github.sef24sp4.core.interfaces;
 
 
 /**
